+++ src/main/java/bolt/packets/ControlPacket.java
@@ -126,7 +126,7 @@ public abstract class ControlPacket implements BoltPacket {
         return destinationID;
     }
 
-    public void setDestinationID(long destinationID) {
+    public void setDestinationID(int destinationID) {
         this.destinationID = destinationID;
     }
